bugfix: extract client-side code to seperate function
